Phase 3 Mega Commit
Commiting all the work for Phase 3 (Inserts). Potential additions in the coming weekend include Updates (Bonus), Heroku application deployment(bonus), and a potential minimal UI(Big if). There is also an unwaranted and newly added error to a method of Phase 2 that will be fixed post haste. And a potential qol addition for a one do all insert, and a auto updating resouce quantity upon purchase/reservation within the orders table.