Final Fixes. All Gets work.
What's missing is the organization of the main (shall be done between 19th and 20th), the double (and triple) verification of all the working methods, gets and posts alike. Potential Update completion(20th to 22nd?), and Heroku implementation (potentially for the 19th or 20th).

If time allows it, commencing the creation of a basic UI is in place, but likely put into last place, as the work required to connect UI's is painful at best.